requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
cloudinary
gunicorn
python-dotenv
